requirements.txt
-i https://pypi.org/simple
numpy==1.19.5
packaging==20.8
pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyyaml==5.3.1
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
jinja2==3.0.1
Bokeh==2.2.2
streamlit==0.80.0
wordcloud==1.8.1
